+++ Fabric/changelog.md
@@ -1,1 +1,1 @@
-Port Fabric integration to 1.19.3
+Port Fabric integration to 1.20.1